ability to change index to hash
