--- db/migrate/20230724121448_create_solid_cache_entries.rb
@@ -1,11 +1,0 @@
-class CreateSolidCacheEntries < ActiveRecord::Migration[7.0]
-  def change
-    create_table :solid_cache_entries do |t|
-      t.binary   :key,        null: false,   limit: 1024
-      t.binary   :value,      null: false,   limit: 512.megabytes
-      t.datetime :created_at, null: false
-
-      t.index    :key,        unique: true
-    end
-  end
-end
